Code cleanup: eliminate partial classes, add RecordHandlerBase, split large files, reorganize folders
Phases 0-5 of structured cleanup:
- Phase 0: Add Roslyn/SonarQube analyzer rules to .editorconfig
- Phase 1: Eliminate non-GUI partial classes (SubrecordSchemaRegistry, FaceGenControls,
  RuntimeBufferAnalyzer, NIF expression nodes, merged small partials)
- Phase 2: Introduce RecordHandlerBase abstract class for 27 ESM record handlers
- Phase 3: Split oversized files (NpcBodyBuilder 1867→213L, ShowCommand 1224→147L,
  RuntimeRefrReader 1201→510L, RuntimeCellReader 1155→646L, NpcExportSceneBuilder 1337→436L)
- Phase 4: Reorganize CLI (60 files into Commands/{Esm,Bsa,Dmp,Dialogue,Save,Version,
  Analysis,Export}/, Shared/, Show/) and Runtime Readers (45 files into Specialized/,
  Generic/, Probes/, Layouts/, Scanning/)
- Phase 5: Extract shared CliProgressRunner and CliTableBuilder utilities,
  deduplicate Spectre.Console progress patterns across 8 commands

280 files changed, ~14K insertions, ~10K deletions. All 1,419 tests pass.

+++ src/FalloutXbox360Utils/Core/Formats/Nif/Rendering/FaceGenTextureMorpher.cs
@@ -344,9 +344,12 @@ internal static class FaceGenTextureMorpher
         {
             TextureAccumulationMode.CurrentFloat => AccumulateNativeDeltasFloat(egt, textureCoeffs),
             TextureAccumulationMode.EngineQuantized256 => AccumulateNativeDeltasQuantized256(egt, textureCoeffs),
-            TextureAccumulationMode.EngineQuantized256Double => AccumulateNativeDeltasQuantized256Double(egt, textureCoeffs),
-            TextureAccumulationMode.EngineQuantizedCombined256 => AccumulateNativeDeltasQuantizedCombined(egt, textureCoeffs, 256),
-            TextureAccumulationMode.EngineQuantizedCombined65536 => AccumulateNativeDeltasQuantizedCombined(egt, textureCoeffs, 65536),
+            TextureAccumulationMode.EngineQuantized256Double => AccumulateNativeDeltasQuantized256Double(egt,
+                textureCoeffs),
+            TextureAccumulationMode.EngineQuantizedCombined256 => AccumulateNativeDeltasQuantizedCombined(egt,
+                textureCoeffs, 256),
+            TextureAccumulationMode.EngineQuantizedCombined65536 => AccumulateNativeDeltasQuantizedCombined(egt,
+                textureCoeffs, 65536),
             _ => throw new ArgumentOutOfRangeException(nameof(accumulationMode), accumulationMode, null)
         };
     }
@@ -442,14 +445,14 @@ internal static class FaceGenTextureMorpher
         var count = Math.Min(textureCoeffs.Length, egt.SymmetricMorphs.Length);
         for (var morphIndex = 0; morphIndex < count; morphIndex++)
         {
-            var coeff256 = (int)((double)textureCoeffs[morphIndex] * 256.0);
+            var coeff256 = (int)(textureCoeffs[morphIndex] * 256.0);
             if (coeff256 == 0)
             {
                 continue;
             }
 
             var morph = egt.SymmetricMorphs[morphIndex];
-            var scale256 = (int)((double)morph.Scale * 256.0);
+            var scale256 = (int)(morph.Scale * 256.0);
             if (scale256 == 0)
             {
                 continue;
